requirements.txt
asgiref==3.7.2
attrs==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.14.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.2.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.2
pytest-django==4.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0